// src/core/enums/permisos.ts
export enum PermisosE {
  LISTAR_GASTO = "LISTAR_GASTO",
  CREAR_GASTO = "CREAR_GASTO",
  EDITAR_GASTO = "EDITAR_GASTO",
  ELIMINAR_GASTO = "ELIMINAR_GASTO",

  CREAR_LECTURA = "CREAR_LECTURA",
  LISTAR_LECTURA = "LISTAR_LECTURA",
  EDITAR_LECTURA = "EDITAR_LECTURA",
  ELIMINAR_LECTURA = "ELIMINAR_LECTURA",

  LISTAR_PAGO = "LISTAR_PAGO",
  CREAR_PAGO = "CREAR_PAGO",
  ANULAR_PAGO = "ANULAR_PAGO",

  LISTAR_TARIFA = "LISTAR_TARIFA",
  EDITAR_TARIFA = "EDITAR_TARIFA",
  CREAR_TARIFA = "CREAR_TARIFA",
  ELIMINAR_TARIFA = "ELIMINAR_TARIFA",

  LISTAR_MEDIDOR = "LISTAR_MEDIDOR",
  EDITAR_MEDIDOR = "EDITAR_MEDIDOR",
  CREAR_MEDIDOR = "CREAR_MEDIDOR",
  ELIMINAR_MEDIDOR = "ELIMINAR_MEDIDOR",

  LISTAR_ROL = "LISTAR_ROL",
  EDITAR_ROL = "EDITAR_ROL",
  CREAR_ROL = "CREAR_ROL",
  ELIMINAR_ROL = "ELIMINAR_ROL",

  LISTAR_USUARIO = "LISTAR_USUARIO",
  EDITAR_USUARIO = "EDITAR_USUARIO",
  CREAR_USUARIO = "CREAR_USUARIO",
  ELIMINAR_USUARIO = "ELIMINAR_USUARIO",

  LISTAR_CLIENTE = "LISTAR_CLIENTE",
  EDITAR_CLIENTE = "EDITAR_CLIENTE",
  CREAR_CLIENTE = "CREAR_CLIENTE",
  ELIMINAR_CLIENTE = "ELIMINAR_CLIENTE",
}